New Skills Checklist Page
Updated the search results to now display a link to a skills checklist for each result.  It will pass to the skills.php page using GET and then is setup to be ready to search the skills table for information to populate the table.  Need to wait to talk with group about connecting to that table and how to handle information on it, and how information is stored in that table for multiple skills.