Continued Skills.php
Continued working on Skills.php to make the skills checklist work.  Setup the basic layout of the page (still needs work), and added the buttons to the page.  They link to a js file but currently do not do anything.
Also in this push, fixed the form-action on index.php of the Search page.